migrate to integeter value

+++ app/Http/Controllers/DashboardController.php
@@ -26,7 +26,7 @@ class DashboardController extends Controller
         ];
 
         if ($authorization_level == 1) {
-            $response['data']['total_asset'] = Transaction::where('id_user', $user->id)
+            $response['data']['total_asset'] = (int) Transaction::where('id_user', $user->id)
                 ->where('status', 'APPROVED')
                 ->sum('investor_amount') ?? 0;
             $response['data']['campaigns'] = Campaign::where('status', 'ACTIVE')->limit(5)->get();
